# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Descripción de este servidor backend",
        "title": "API Backend de la aplicación besigabi",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1/",
    "paths": {
...
        "/api/v1/besigabi/skl/consumirCredenciales": {
            "post": {
                "description": "Valida las credenciales de un usuario",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SKL"
                ],
                "summary": "Consumir credenciales",
                "parameters": [
                    {
                        "description": "Credenciales",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_SKL_domain.ConsumirCredencialesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_SKL_domain.ConsumirCredencialesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/besigabi/skl/sedesUsuario": {
            "post": {
                "description": "Consulta las sedes disponibles para un usuario",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SKL"
                ],
                "summary": "Obtener sedes por usuario",
                "parameters": [
                    {
                        "description": "Parámetros de usuario",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_SKL_domain.SedesUsuarioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_SKL_domain.SedeUsuarioResponse"
                            }
                        }
                    },
                    "204": {
                        "description": "Sin sedes"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/besigabi/skl/serviciosSiga": {
            "post": {
                "description": "Consulta los servicios SIGA disponibles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SKL"
                ],
                "summary": "Obtener servicios SIGA",
                "parameters": [
                    {
                        "description": "Parámetros de la consulta de servicios",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_SKL_domain.ServiciosSigaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_SKL_domain.ServicioSigaResponse"
                            }
                        }
                    },
                    "204": {
                        "description": "Sin servicios"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/besigabi/skl/taquillas": {
            "post": {
                "description": "Consulta las taquillas activas para una sede y módulo específico",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Taquillas"
                ],
                "summary": "Obtener taquillas por sede y módulo",
                "parameters": [
                    {
                        "description": "Parámetros de la consulta de taquillas",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_SKL_domain.TaquillasRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_SKL_domain.TaquillaResponse"
                            }
                        }
                    },
                    "204": {
                        "description": "Sin taquillas"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/besigabi/skl/turnosDisponibles": {
            "post": {
                "description": "Consulta turnos disponibles asociados a una orden",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SKL"
                ],
                "summary": "Obtener turnos disponibles con orden",
                "parameters": [
                    {
                        "description": "Parámetros de la consulta",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_SKL_domain.TurnosDisponiblesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_SKL_domain.TurnoDisponibleResponse"
                            }
                        }
                    },
                    "204": {
                        "description": "Sin turnos"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/besigabi/skl/turnosDisponiblesLis": {
            "post": {
                "description": "Consulta los turnos disponibles en LIS",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SKL"
                ],
                "summary": "Obtener turnos disponibles",
                "parameters": [
                    {
                        "description": "Parámetros de la consulta",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_SKL_domain.TurnosDisponiblesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_SKL_domain.TurnoDisponibleResponse"
                            }
                        }
                    },
                    "204": {
                        "description": "Sin turnos"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "develop_private_CLTech_besigabi_internal_api_SKL_domain.ConsumirCredencialesRequest": {
            "type": "object",
            "properties": {
                "contrasena": {
                    "type": "string"
                },
                "usuario": {
                    "type": "string"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_SKL_domain.ConsumirCredencialesResponse": {
            "type": "object",
            "properties": {
                "idUsuario": {
                    "type": "integer"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_SKL_domain.PacienteTurno": {
            "type": "object",
            "properties": {
                "apellido1": {
                    "type": "string"
                },
                "apellido2": {
                    "type": "string"
                },
                "fechaCumple": {
                    "type": "string"
                },
                "nomTipoDocumento": {
                    "type": "string"
                },
                "nombre1": {
                    "type": "string"
                },
                "nombre2": {
                    "type": "string"
                },
                "numeroDocumento": {
                    "type": "string"
                },
                "sexo": {
                    "type": "integer"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_SKL_domain.SedeUsuarioResponse": {
            "type": "object",
            "properties": {
                "codSede": {
                    "type": "string"
                },
                "idSede": {
                    "type": "integer"
                },
                "nomSede": {
                    "type": "string"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_SKL_domain.SedesUsuarioRequest": {
            "type": "object",
            "properties": {
                "usuario": {
                    "type": "string"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_SKL_domain.ServicioSigaResponse": {
            "type": "object",
            "properties": {
                "codServicio": {
                    "type": "string",
                    "example": "SV01"
                },
                "idServicio": {
                    "type": "integer",
                    "example": 1
                },
                "nomServicio": {
                    "type": "string",
                    "example": "Servicio 1"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_SKL_domain.ServiciosSigaRequest": {
            "type": "object",
            "properties": {
                "codModulo": {
                    "type": "integer",
                    "example": 4000
                },
                "idSede": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_SKL_domain.TaquillaResponse": {
            "type": "object",
            "properties": {
                "codTaquilla": {
                    "type": "string",
                    "example": "TQ01"
                },
                "estadoTaquilla": {
                    "type": "integer",
                    "example": 0
                },
                "idTaquilla": {
                    "type": "integer",
                    "example": 1
                },
                "nIdModulo": {
                    "type": "integer",
                    "example": 4000
                },
                "nomTaquilla": {
                    "type": "string",
                    "example": "Taquilla 1"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_SKL_domain.TaquillasRequest": {
            "type": "object",
            "properties": {
                "codModulo": {
                    "type": "integer",
                    "example": 4000
                },
                "idSede": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_SKL_domain.TurnoDisponibleResponse": {
            "type": "object",
            "properties": {
                "codEmpresa": {
                    "type": "string"
                },
                "codServicio": {
                    "type": "string"
                },
                "codTipoTurno": {
                    "type": "string"
                },
                "fechaHoraTurno": {
                    "type": "string"
                },
                "idEmpresa": {
                    "type": "integer"
                },
                "idSede": {
                    "type": "integer"
                },
                "idServicio": {
                    "type": "integer"
                },
                "idTipoDocumento": {
                    "type": "integer"
                },
                "idTipoTurno": {
                    "type": "integer"
                },
                "minutosEspera": {
                    "type": "integer"
                },
                "nPrioridad": {
                    "type": "integer"
                },
                "nomEmpresa": {
                    "type": "string"
                },
                "nomServicio": {
                    "type": "string"
                },
                "nomTipoTurno": {
                    "type": "string"
                },
                "nombreSede": {
                    "type": "string"
                },
                "numeroOrden": {
                    "type": "string"
                },
                "numeroTurno": {
                    "type": "string"
                },
                "paciente": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_SKL_domain.PacienteTurno"
                },
                "tiempo": {
                    "type": "integer"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_SKL_domain.TurnosDisponiblesRequest": {
            "type": "object",
            "properties": {
                "idSede": {
                    "type": "integer"
                },
                "idServicio": {
                    "type": "integer"
                },
                "idTaquilla": {
                    "type": "integer"
                }
            }
        },
...
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
    }
}